♻️ refactor: 적립 포인트 로직을 별도 컴포넌트로 분리
- RightColumn과 main.basic.js에서 새로운 컴포넌트를 사용하도록 수정
- 기존의 인라인 계산 및 렌더링 로직 제거
